Fix bugs on tool-component, Add csrf token to mail post,  add cloning repo feature
