+++ composer.json
@@ -10,7 +10,6 @@
     "require": {
         "php": "^7.4",
         "ext-json": "*",
-        "cpliakas/git-wrapper": "^3.0",
         "fideloper/proxy": "^4.2",
         "friendsofphp/php-cs-fixer": "^2.16",
         "fruitcake/laravel-cors": "^1.0",
