+++ routes/web.php
@@ -21,7 +21,6 @@ Route::get('/', function () {
 //Route::post('/test', 'TestController@testAxiosRouting');
 Route::get('/test', 'ReportController@test');
 Route::post('/test', 'ReportController@test');
-Route::get('/exist', 'ReportController@isRepo');
-Route::post('/getRepos','ReportController@getUserRepositories');//TODO create Repo controller or User controller
-Route::get('/cloneRepo','ReportController@getUserRepository');//TODO create Repo controller or User controller
+Route::post('/getRepos', 'GitController@getUserRepositories');
+Route::post('/cloneRepo', 'GitController@cloneRepository');
 Route::post('/mail', 'ReportController@mail');
